Verify NDPProxyInterface exists before loading
The bridge_store needs to check if the NDPProxyInterface value was stored before
loading its value.

From aboch:
Given this new field is not there for existing bridge network in store, you need to
first check if the entry exists in the map before converting to string, or it will
panic.

Signed-off-by: Zvi "CtrlZvi" Effron <viz+GitHub@flippedperspective.com>